Allow handling tabs shouldSelectViewController
When a tab is selected, iOS first calls the method
'tabBarController:shouldSelectViewController' to determine whether
or not it should actually select the tab. Returning false prevents
the tab from being selected.

This commit enables the tab delegate to implement this functionality.
The method is called on the cocoatouch event but also when `open_tab`
is called programmatically to ensure consistent behavior. If the
delegate does not implement a handler the method just returns true
so this change is backwards compatibile.